Moving parser tests to a better structure
and covering a bit more parts
